+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.9.1] - 2026-04-21
+
+### Fixed
+
+- derived state in notification pages
+
 ## [0.9.0] - 2026-04-21
 
 ### Added